Rename NoRunCliktCommand to NoOpCliktCommand
